Move name/line_breaks check to googlefonts profile
as it is a vendor-specific policy rather than an opentype spec requirement.
(issue #2778)
